+++ resources/js/Pages/Contact.vue
@@ -60,7 +60,10 @@
             <label for="message" class="block text-gray-200 text-sm font-bold mb-2">
               Message:
             </label>
-            <textarea id="message" v-model="form.message" rows="4" class="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+            <textarea id="message"
+                      v-model="form.message"
+                      rows="4"
+                      class="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
                       required></textarea>
             <div v-if="form.errors.message" v-text="form.errors.message"
                  class="text-xs text-red-600 mt-1"></div>
@@ -127,7 +130,7 @@ let form = useForm({
 })
 
 let submit = () => {
-  form.post(route('public.contact'))
+  form.post(route('public.contact.submit'))
 }
 
 const marginTopClass = computed(() => {