+++ resources/js/Stores/AppSettingStore.js
@@ -15,6 +15,7 @@ const initialState = () => ({
     currentPage: '', // formerly videoPlayerStore.currentPage
     fullPage: false, // Used to determine layout FullPage or TopRight
     pageIsHidden: true, // Used to hide the page when fullPage = false && showOtt = true
+    shouldScrollToTop: false,
     ott: 0, // Number representing the Ott Panel currently open. 0 is closed.
             // 1: Info
             // 2: Channels