+++ resources/js/Stores/UserStore.js
@@ -206,7 +206,20 @@ export const useUserStore = defineStore('userStore', {
                 console.error('Timezone is not set.')
                 return dateTime // Or handle this case as appropriate for your app
             }
+            if (!dateTime) {
+                return ''; // Return an empty string for invalid or null dates
+            }
             return dayjs.utc(dateTime).tz(this.timezone).format(formatString)
+        },
+        formatDateToISOInUserTimezone(dateTime) {
+            if (!this.timezone) {
+                console.error('Timezone is not set.')
+                return dateTime // Or handle this case as appropriate for your app
+            }
+            if (!dateTime) {
+                return ''; // Return an empty string for invalid or null dates
+            }
+            return dayjs.utc(dateTime).tz(this.timezone).format('YYYY-MM-DD')
         },
         async setUserTimezone(newTimezone) {
             this.timezone = newTimezone